requirements.txt
langchain>=0.3.26,<0.4
langchain-community>=0.3.26,<0.4
langchain-pinecone>=0.2.8
langchain-groq>=0.1.10

flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
groq>=0.37.0

-e .
